Datcore 196 make confighandler skip if none

+++ hamb/handlers/email_handler.py
@@ -30,48 +30,49 @@ class Handler(object):
         :param conf:
         :return:
         """
-        recipients = conf.split(" ")
+        if conf:
+            recipients = conf.split(" ")
 
-        level = result["summary"]["status"]
-        manifest = result["summary"]["manifest"]
-        subject = (
-            str(self.environment).upper()
-            + str(" Hambot %s: %s" % (level, manifest)).title()
-        )
-
-        with_attachment = os.path.exists("diagnostic_query_results.csv")
-
-        if level == "success" or not with_attachment:
-            json_msg = json.dumps(result, indent=4, default=json_serializer)
-            if self.environment != "dev":
-                Email.send_mail(
-                    aws_access_key=self.aws_id,
-                    aws_secret_key=self.aws_key,
-                    aws_sender=self.ses_def_sender,
-                    aws_region=self.ses_region,
-                    to_addr=recipients,
-                    subject=subject,
-                    text_msg=json_msg,
-                )
-
-        else:
-            json_msg = (
-                json.dumps(result, indent=4, default=json_serializer)
-                .replace(" ", "&nbsp;")
-                .replace("\n", "<br>")
+            level = result["summary"]["status"]
+            manifest = result["summary"]["manifest"]
+            subject = (
+                str(self.environment).upper()
+                + str(" Hambot %s: %s" % (level, manifest)).title()
             )
-            if self.environment != "dev":
-                Email.send_email_with_attachment(
-                    aws_access_key=self.aws_id,
-                    aws_secret_key=self.aws_key,
-                    aws_sender=self.ses_def_sender,
-                    aws_region=self.ses_region,
-                    to_addr=recipients,
-                    subject=subject,
-                    body_msg=json_msg,
-                    filename="diagnostic_query_results.csv",
+
+            with_attachment = os.path.exists("diagnostic_query_results.csv")
+
+            if level == "success" or not with_attachment:
+                json_msg = json.dumps(result, indent=4, default=json_serializer)
+                if self.environment != "dev":
+                    Email.send_mail(
+                        aws_access_key=self.aws_id,
+                        aws_secret_key=self.aws_key,
+                        aws_sender=self.ses_def_sender,
+                        aws_region=self.ses_region,
+                        to_addr=recipients,
+                        subject=subject,
+                        text_msg=json_msg,
+                    )
+
+            else:
+                json_msg = (
+                    json.dumps(result, indent=4, default=json_serializer)
+                    .replace(" ", "&nbsp;")
+                    .replace("\n", "<br>")
                 )
-            self.os.remove("diagnostic_query_results.csv")
+                if self.environment != "dev":
+                    Email.send_email_with_attachment(
+                        aws_access_key=self.aws_id,
+                        aws_secret_key=self.aws_key,
+                        aws_sender=self.ses_def_sender,
+                        aws_region=self.ses_region,
+                        to_addr=recipients,
+                        subject=subject,
+                        body_msg=json_msg,
+                        filename="diagnostic_query_results.csv",
+                    )
+                self.os.remove("diagnostic_query_results.csv")
 
 
 def render_html(result):
